pyproject.toml
[project]
name = "api-testing-project"
version = "0.1.0"
description = "API tests for Practice ExpandTesting API"
authors = [{name = "Tester", email = "tester@example.com"}]
requires-python = ">=3.13"

dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "pydantic[email]>=2.0.0",
    "allure-pytest>=2.13.0",
    "python-dotenv>=1.0.0",
    "pysonar>=1.2.1.3951",
    "faker>=38.2.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "smoke: marks tests as smoke tests",
    "regression: marks tests as regression tests",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_unimported = true
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
show_error_codes = true
cache_dir = ".venv/.cache/mypy_cache"
exclude = ["^\\.venv/"]

[tool.ruff]
target-version = "py313"
line-length = 100
indent-width = 4
src = ["src"]

[tool.ruff.lint]
preview = true
select = [
    "UP",  # pyupgrade
    "I",  # isort
    "E",  # flake8-errors
    "E4",
    "E7",
    "E9",
    "F",
    "W",  # flake8-warnings
    "F",  # flake8-pyflakes
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "S",  # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",  # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB",  # refurb
]
ignore = [
    "C408",  # Ненужный вызов dict, list и т.д.
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "RUF001",  # Игнорирование похожих символов с разными ASCII
    "RUF002",  # Игнорирование похожих символов с разными ASCII
    "RUF003",  # Игнорирование похожих символов с разными ASCII
    "S110",  # try-except-pass
    "E501",  # line-too-long
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "tests/**",
]
fixable = [
    "I",  # сортировка импортов
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.flake8-bandit]
check-typed-exception = true

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
